requirements.txt
anthropic>=0.18.0
openpyxl>=3.1.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
pydantic>=2.0.0
streamlit>=1.28.0
